Re-enable, add and fix tests for Templates / Refactor Template.js Class (#1292)
* Re-enable unit tests for templates
* Remove appsody, refactor getTemplates
* Complete unit tests (initializeRepositoryList was missing)
* Update open API docs and refactor test

+++ test/modules/template.service.js
@@ -78,161 +78,6 @@ const defaultCodewindTemplates = [
     },
 ];
 
-const defaultAppsodyTemplates = [
-    {
-        label: 'Appsody Eclipse MicroProfile® template',
-        description: 'Eclipse MicroProfile on Open Liberty & OpenJ9 using Maven',
-        language: 'java',
-        url: 'https://github.com/appsody/stacks/releases/download/java-microprofile-v0.2.18/incubator.java-microprofile.v0.2.18.templates.default.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-    {
-        label: 'Appsody LoopBack 4 template',
-        description: 'LoopBack 4 API Framework for Node.js',
-        language: 'nodejs',
-        url: 'https://github.com/appsody/stacks/releases/download/nodejs-loopback-v0.1.5/incubator.nodejs-loopback.v0.1.5.templates.scaffold.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-    {
-        label: 'Appsody Node.js Express scaffold template',
-        description: 'Express web framework for Node.js',
-        language: 'nodejs',
-        url: 'https://github.com/appsody/stacks/releases/download/nodejs-express-v0.2.7/incubator.nodejs-express.v0.2.7.templates.scaffold.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-    {
-        label: 'Appsody Node.js Express simple template',
-        description: 'Express web framework for Node.js',
-        language: 'nodejs',
-        url: 'https://github.com/appsody/stacks/releases/download/nodejs-express-v0.2.7/incubator.nodejs-express.v0.2.7.templates.simple.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-    {
-        label: 'Appsody Node.js template',
-        description: 'Runtime for Node.js applications',
-        language: 'nodejs',
-        url: 'https://github.com/appsody/stacks/releases/download/nodejs-v0.2.5/incubator.nodejs.templates.simple.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-    {
-        label: 'Appsody Python Flask template',
-        description: 'Flask web Framework for Python',
-        language: 'python',
-        url: 'https://github.com/appsody/stacks/releases/download/python-flask-v0.1.4/incubator.python-flask.v0.1.4.templates.simple.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-    {
-        label: 'Appsody Spring Boot® default template',
-        description: 'Spring Boot using OpenJ9 and Maven',
-        language: 'java',
-        url: 'https://github.com/appsody/stacks/releases/download/java-spring-boot2-v0.3.15/incubator.java-spring-boot2.v0.3.15.templates.default.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-    {
-        label: 'Appsody Spring Boot® kotlin template',
-        description: 'Spring Boot using OpenJ9 and Maven',
-        language: 'java',
-        url: 'https://github.com/appsody/stacks/releases/download/java-spring-boot2-v0.3.15/incubator.java-spring-boot2.v0.3.15.templates.kotlin.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-    {
-        label: 'Appsody starter sample runnable stack template',
-        description: 'starter sample stack that can be run by ' +
-          'appsody, to help creation of more stacks',
-        language: 'bash',
-        url: 'https://github.com/appsody/stacks/releases/download/starter-v0.1.0/incubator.starter.v0.1.0.templates.simple.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-    {
-        label: 'Appsody Swift template',
-        description: 'Runtime for Swift applications',
-        language: 'swift',
-        url: 'https://github.com/appsody/stacks/releases/download/swift-v0.1.4/incubator.swift.templates.simple.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-    },
-];
-
-const defaultKabaneroTemplates = [
-    {
-        label: 'Kabanero Eclipse MicroProfile® template',
-        description: 'Eclipse MicroProfile on Open Liberty & OpenJ9 using Maven',
-        language: 'java',
-        url: 'https://github.com/kabanero-io/collections/releases/download/0.2.1/incubator.java-microprofile.v0.2.18.templates.default.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-        source: 'Kabanero Collections',
-    },
-    {
-        label: 'Kabanero LoopBack 4 template',
-        description: 'LoopBack 4 API Framework for Node.js',
-        language: 'nodejs',
-        url: 'https://github.com/kabanero-io/collections/releases/download/0.2.1/incubator.nodejs-loopback.v0.1.5.templates.scaffold.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-        source: 'Kabanero Collections',
-    },
-    {
-        label: 'Kabanero Node.js Express simple template',
-        description: 'Express web framework for Node.js',
-        language: 'nodejs',
-        url: 'https://github.com/kabanero-io/collections/releases/download/0.2.1/incubator.nodejs-express.v0.2.5.templates.simple.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-        source: 'Kabanero Collections',
-    },
-    {
-        label: 'Kabanero Node.js Express skaffold template',
-        description: 'Express web framework for Node.js',
-        language: 'nodejs',
-        url: 'https://github.com/kabanero-io/collections/releases/download/0.2.1/incubator.nodejs-express.v0.2.7.templates.scaffold.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-        source: 'Kabanero Collections',
-    },
-    {
-        label: 'Kabanero Node.js template',
-        description: 'Runtime for Node.js applications',
-        language: 'nodejs',
-        url: 'https://github.com/kabanero-io/collections/releases/download/0.2.1/incubator.nodejs.v0.2.5.templates.simple.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-        source: 'Kabanero Collections',
-    },
-    {
-        label: 'Kabanero Spring Boot® default template',
-        description: 'Spring Boot using OpenJ9 and Maven',
-        language: 'java',
-        url: 'https://github.com/kabanero-io/collections/releases/download/0.2.1/incubator.java-spring-boot2.v0.3.15.templates.default.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-        source: 'Kabanero Collections',
-    },
-    {
-        label: 'Kabanero Spring Boot® kotlin template',
-        description: 'Spring Boot using OpenJ9 and Maven',
-        language: 'java',
-        url: 'https://github.com/kabanero-io/collections/releases/download/0.2.1/incubator.java-spring-boot2.v0.3.15.templates.kotlin.tar.gz',
-        projectType: 'appsodyExtension',
-        projectStyle: 'Appsody',
-        source: 'Kabanero Collections',
-    },
-];
-
-const defaultTemplates = [
-    ...defaultCodewindTemplates,
-    ...defaultAppsodyTemplates,
-];
-
 const styledTemplates = {
     codewind: {
         label: 'Codewind template',
@@ -262,24 +107,7 @@ const sampleRepos = {
         projectStyles: ['Codewind'],
         name: 'Default templates',
     },
-    appsody: {
-        url: 'https://raw.githubusercontent.com/kabanero-io/codewind-appsody-templates/master/devfiles/index.json',
-        description: 'Appsody extension for Codewind',
-        enabled: true,
-        projectStyles: ['Appsody'],
-    },
-    fromAppsodyExtension: {
-        url: 'https://github.com/appsody/stacks/releases/latest/download/incubator-index.json',
-        description: 'Appsody Stacks - appsodyhub',
-        enabled: true,
-        protected: true,
-        projectStyles: ['Appsody'],
-    },
 };
-const defaultRepoList = [
-    sampleRepos.codewind,
-    sampleRepos.fromAppsodyExtension,
-];
 
 const validUrlNotPointingToIndexJson = 'https://support.oneskyapp.com/hc/en-us/article_attachments/202761627/example_1.json';
 
@@ -352,9 +180,11 @@ async function getTemplates(queryParams) {
  */
 async function setTemplateReposTo(repoList) {
     const reposToDelete = (await getTemplateRepos()).body;
-    await Promise.all(reposToDelete.map(repo =>
-        deleteTemplateRepo(repo.url)
-    ));
+    if (reposToDelete.length > 0) {
+        await Promise.all(reposToDelete.map(repo =>
+            deleteTemplateRepo(repo.url)
+        ));
+    }
     await Promise.all(repoList.map(repo =>
         addTemplateRepo(repo)
     ));
@@ -394,12 +224,9 @@ function saveReposBeforeEachTestAndRestoreAfterEach() {
 }
 
 module.exports = {
-    defaultTemplates,
     defaultCodewindTemplates,
-    defaultKabaneroTemplates,
     styledTemplates,
     sampleRepos,
-    defaultRepoList,
     validUrlNotPointingToIndexJson,
     getTemplateRepos,
     addTemplateRepo,
